Changes to worm angle

+++ common_src/snapshot.h
@@ -44,7 +44,7 @@ public:
     char id;
     float pos_x;
     float pos_y;
-    float angle;
+    int angle;
     int max_health;
     int health;
     char direction;
@@ -52,7 +52,7 @@ public:
     int state;
 
 
-    WormSnapshot(char id, float pos_x, float pos_y, float angle, int max_health, int health, char direction, int weapon,
+    WormSnapshot(char id, float pos_x, float pos_y, int angle, int max_health, int health, char direction, int weapon,
                  int state):
             id(id),
             pos_x(pos_x),
